+++ frontend/src/features/coordinator-dashboard/index.ts
@@ -1,1 +1,2 @@
 export {GuestInviteButton} from './GuestInviteButton';
+export {LoadingComponent} from './LoadingComponent';
