+++ frontend/src/pages/authentication/EmailVerificationSuccess.tsx
