+++ frontend/src/pages/authentication/ForgotPassword.tsx
@@ -11,10 +11,10 @@ import CloseIcon from '@mui/icons-material/Close';
 import {useFormikContext} from 'formik';
 import React from 'react';
 import {useNavigate} from 'react-router-dom';
-import {isErrorWithMessage, isFetchBaseQueryError} from '../redux/helpers';
-import {ResestPasswordValues} from '../features/authentication/ResetPasswordContext';
-import {useForgotPasswordMutation} from '../services/auth';
-import {FormContainer} from '../features/authentication';
+import {isErrorWithMessage, isFetchBaseQueryError} from '../../redux/helpers';
+import {ResestPasswordValues} from '../../features/authentication/ResetPasswordContext';
+import {useForgotPasswordMutation} from '../../services/auth';
+import {FormContainer} from '../../features/authentication';
 
 export const ForgotPassword = () => {
   const [errorMessage, setErrorMessage] = React.useState('');
