+++ frontend/src/pages/authentication/ForgotPasswordCode.tsx
@@ -13,10 +13,10 @@ import CloseIcon from '@mui/icons-material/Close';
 import {useFormikContext} from 'formik';
 import React from 'react';
 import {useNavigate} from 'react-router-dom';
-import {ResestPasswordValues} from '../features/authentication/ResetPasswordContext';
-import {CodeField, FormContainer} from '../features/authentication';
-import {useForgotPasswordMutation} from '../services/auth';
-import {getErrorMessage} from '../redux/helpers';
+import {ResestPasswordValues} from '../../features/authentication/ResetPasswordContext';
+import {CodeField, FormContainer} from '../../features/authentication';
+import {useForgotPasswordMutation} from '../../services/auth';
+import {getErrorMessage} from '../../redux/helpers';
 
 interface Alert {
   severity: AlertColor;
