+++ frontend/src/pages/authentication/SelectAccountType.tsx
@@ -2,7 +2,7 @@ import React, {useState} from 'react';
 import {useNavigate} from 'react-router-dom';
 import {Typography, Button, Stack, Link, Divider} from '@mui/material';
 import {styled} from '@mui/system';
-import {HostIcon, CoordinatorIcon} from '../features/common/icons';
+import {HostIcon, CoordinatorIcon} from '../../features/common/icons';
 
 export const SelectAccountType = () => {
   const [type, setType] = useState('');
