+++ frontend/src/pages/index.ts
@@ -8,16 +8,16 @@ export {GuestDocuments} from './GuestDocuments';
 export {Home} from './Home';
 export {HostApplicationTracker} from './HostApplicationTracker';
 export {HostsList} from './HostsList';
-export {ForgotPassword} from './ForgotPassword';
-export {ResetPassword} from './ResetPassword';
-export {SelectAccountType} from './SelectAccountType';
-export {SignIn} from './SignIn';
-export {SignUp} from './SignUp';
-export {EmailVerificationSuccess} from './EmailVerificationSuccess';
-export {EmailVerificationError} from './EmailVerificationError';
+export {ForgotPassword} from './authentication/ForgotPassword';
+export {ResetPassword} from './authentication/ResetPassword';
+export {SelectAccountType} from './authentication/SelectAccountType';
+export {SignIn} from './authentication/SignIn';
+export {SignUp} from './authentication/SignUp';
+export {EmailVerificationSuccess} from './authentication/EmailVerificationSuccess';
+export {EmailVerificationError} from './authentication/EmailVerificationError';
 export {NewPassword} from './NewPassword';
-export {ForgotPasswordCode} from './ForgotPasswordCode';
-export {ForgotPasswordSuccess} from './ForgotPasswordSuccess';
+export {ForgotPasswordCode} from './authentication/ForgotPasswordCode';
+export {ForgotPasswordSuccess} from './authentication/ForgotPasswordSuccess';
 export {Settings} from './Settings';
-export {ConfirmSignUp} from './ConfirmSignUp';
+export {ConfirmSignUp} from './authentication/ConfirmSignUp';
 export {IntakeProfile} from './IntakeProfile';
